Criar Arquivo na Maquina

+++ APIGerarBoletos.Core/Services/GeradorItauV2.cs
@@ -128,15 +128,23 @@ namespace APIGerarBoletos.Services
             }
         }
 
-        private void SaveBoletoPDF(byte[] boletoPdf)
+        private MemoryStream SaveBoletoPDF(byte[] boletoPdf, bool isFile = false)
         {
             MemoryStream stream = new MemoryStream();
             stream.Write(boletoPdf, 0, boletoPdf.Length);
-
-            var fileStream = File.Create("C:\\boleto.pdf");
             stream.Seek(0, SeekOrigin.Begin);
-            stream.CopyTo(fileStream);
-            fileStream.Close();
+
+            if (isFile)
+            {
+                if (File.Exists("C:\\boleto.pdf"))
+                    File.Delete("C:\\boleto.pdf");
+
+                var fileStream = File.Create("C:\\boleto.pdf");
+                stream.CopyTo(fileStream);
+                fileStream.Close();
+            }
+
+            return stream;
         }
     }
 }
